Fixing sign extension in divuw and remuw

+++ src/generator/x86_64/helper/interpreter.cpp
@@ -679,7 +679,7 @@ extern "C" uint64_t unresolved_ijump_handler(uint64_t pc) {
             break;
         case FRV_DIVUW:
             if (instr.rd != 0) {
-                register_file[instr.rd] = sign_extend_int64_t(static_cast<uint32_t>(register_file[instr.rs1]) / static_cast<uint32_t>(register_file[instr.rs2]));
+                register_file[instr.rd] = sign_extend_int64_t(static_cast<int32_t>(static_cast<uint32_t>(register_file[instr.rs1]) / static_cast<uint32_t>(register_file[instr.rs2])));
             }
             break;
         case FRV_REM:
@@ -699,7 +699,7 @@ extern "C" uint64_t unresolved_ijump_handler(uint64_t pc) {
             break;
         case FRV_REMUW:
             if (instr.rd != 0) {
-                register_file[instr.rd] = sign_extend_int64_t(static_cast<uint32_t>(register_file[instr.rs1]) % static_cast<uint32_t>(register_file[instr.rs2]));
+                register_file[instr.rd] = sign_extend_int64_t(static_cast<int32_t>(static_cast<uint32_t>(register_file[instr.rs1]) % static_cast<uint32_t>(register_file[instr.rs2])));
             }
             break;
